refactor(tui): extract frosted-glass walls to pixel_painter/glass.rs
Pull the self-contained glass room-divider block (WALL_THICK_*_PX,
GLASS_SEAM_STRIDE, GLASS_CAP_PX, glass_tones, stitch_vertical_wall,
glass_over, paint_glass_wall_h/v) out of pixel_painter/mod.rs into a
focused glass.rs child module. mod.rs 1709 -> 1555 lines.

Behavior-preserving: drawable.rs reaches paint_glass_wall_h via the
unchanged `super::` path (mod.rs re-imports it); the test block resolves
the helpers through `use super::*`. Done before the Rgb sweep so the sweep
edits the relocated file once. tui/CLAUDE.md child-module list updated.
